Adding tags table and utils
